requirements.txt
django-environ 
cookies==2.2.1
coreapi==2.3.3
coreschema==0.0.4  
Django==2.1.7
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-livereload-server==0.3.2
django-model-utils==3.2.0
django-rest-auth==0.9.3
django-rest-swagger==2.1.1
django-filter==2.2.0
djangorestframework==3.9.2
oauthlib==3.1.0
openapi-codec==1.3.2
simplejson==3.11.1
django-background-tasks==1.2.5
django-cors-headers==3.2.1
schedule==0.6.0
psycopg2==2.7.3.2
psycopg2-binary==2.7.7
pygments==2.6.1
